+++ test/moment_dart_test.dart
@@ -28,8 +28,8 @@ void main() {
     final epochPlusFiveDays = epoch + Duration(days: 5);
     final epochPlusAYear = epoch + Duration(days: 365);
 
-    print(moment.format("dd HH:mm, MMMM/DD/YYYY"));
-    print(epoch.format("dd HH:mm, MMMM/DD/YYYY"));
+    print(moment.format(r"dddd HH:mm, MMMM/DD/YYYY"));
+    print(epoch.format(r"dddd HH:mm, MMMM/DD/YYYY"));
 
     expect(moment.lastMonday().weekday, 1);
     expect(localization.relative(const Duration(seconds: 2)), "in a few seconds");