+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+- Fixed `Moment.format` function
+- Fixed negative sign appearing in relative date texts
+
 ## 0.1.0
 
 - Initial beta release.